requirements.txt
Pillow>=9.4.0
pytesseract>=0.3.10
pymupdf>=1.21.1
tqdm>=4.64.1
numpy>=1.24.1